+++ extension-compose/src/main/java/com/mapbox/maps/extension/compose/internal/SettingsUtils.kt
@@ -28,5 +28,6 @@ internal fun GesturesSettingsInterface.applySettings(gesturesSettings: GesturesS
       gesturesSettings.increasePinchToZoomThresholdWhenRotating
     zoomAnimationAmount = gesturesSettings.zoomAnimationAmount
     pinchScrollEnabled = gesturesSettings.pinchScrollEnabled
+    useNativeFlingDeceleration = gesturesSettings.useNativeFlingDeceleration
   }
 }